+++ R/PlotRsmOperatingCharacteristics.R
@@ -28,7 +28,7 @@
 #'    lesion distribution for diseased cases. The first column contains the 
 #'    actual numbers of lesions per case. The second column contains the fraction 
 #'    of diseased cases with the number of lesions specified in the first column. 
-#'    The second column must sum to unity. See \link{UtilLesionDistribution}. 
+#'    The second column must sum to unity. See \link{UtilLesionDistr}. 
 #' 
 #' @param lesWghtDistr The lesion weights distribution, an [1:maxLL,1:maxLL] array. 
 #'    The probability mass function of the